+++ ta4j-core/src/test/java/org/ta4j/core/utils/BarSeriesUtilsTest.java
@@ -193,6 +193,132 @@ public class BarSeriesUtilsTest extends AbstractIndicatorTest<BarSeries, Num> {
         assertEquals(7, this.series.getBarCount());
     }
 
+    /**
+     * Tests if the previous bar is replaced by newBar
+     */
+    @Test
+    public void replaceBarsIfChangedTest() {
+
+        final List<Bar> bars = new ArrayList<>();
+        this.time = Instant.parse("2019-06-01T01:01:00Z");
+
+        final Bar bar0 = new MockBarBuilder(numFactory).endTime(time)
+                .openPrice(1d)
+                .closePrice(2d)
+                .highPrice(3d)
+                .lowPrice(4d)
+                .amount(5d)
+                .volume(0d)
+                .trades(7)
+                .build();
+
+        final Bar bar1 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(1)))
+                .openPrice(1d)
+                .closePrice(1d)
+                .highPrice(1d)
+                .lowPrice(1d)
+                .amount(1d)
+                .volume(1d)
+                .trades(1)
+                .build();
+
+        final Bar bar2 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(2)))
+                .openPrice(2d)
+                .closePrice(2d)
+                .highPrice(2d)
+                .lowPrice(2d)
+                .amount(2d)
+                .volume(2d)
+                .trades(2)
+                .build();
+
+        final Bar bar3 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(3)))
+                .openPrice(3d)
+                .closePrice(3d)
+                .highPrice(3d)
+                .lowPrice(3d)
+                .amount(3d)
+                .volume(3d)
+                .trades(3)
+                .build();
+
+        final Bar bar4 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(4)))
+                .openPrice(3d)
+                .closePrice(4d)
+                .highPrice(5d)
+                .lowPrice(6d)
+                .amount(4d)
+                .volume(4d)
+                .trades(4)
+                .build();
+
+        final Bar bar5 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(5)))
+                .openPrice(4d)
+                .closePrice(5d)
+                .highPrice(5d)
+                .lowPrice(5d)
+                .amount(5d)
+                .volume(5d)
+                .trades(5)
+                .build();
+
+        final Bar bar6 = new MockBarBuilder(numFactory).endTime(time.plus(Duration.ofDays(6)))
+                .openPrice(6d)
+                .closePrice(6d)
+                .highPrice(6d)
+                .lowPrice(6d)
+                .amount(6d)
+                .volume(6d)
+                .trades(6)
+                .build();
+
+        bars.add(bar0);
+        bars.add(bar1);
+        bars.add(bar2);
+        bars.add(bar3);
+        bars.add(bar4);
+        bars.add(bar5);
+        bars.add(bar6);
+
+        this.series = new MockBarSeriesBuilder().withNumFactory(this.numFactory)
+                .withName("Series Name")
+                .withBars(bars)
+                .build();
+
+        final var newBar3 = this.series.barBuilder()
+                .endTime(bar3.getEndTime())
+                .openPrice(1d)
+                .closePrice(1d)
+                .highPrice(1d)
+                .lowPrice(1d)
+                .volume(1d)
+                .trades(33)
+                .build();
+
+        final Bar newBar5 = this.series.barBuilder()
+                .endTime(bar5.getEndTime())
+                .openPrice(1d)
+                .closePrice(1d)
+                .highPrice(1d)
+                .lowPrice(1d)
+                .amount(1d)
+                .volume(1d)
+                .trades(55)
+                .build();
+
+        var newBars = new ArrayList<Bar>();
+        newBars.add(newBar3);
+        newBars.add(newBar5);
+
+        // newBar3 must be replaced with bar3,
+        // newBar5 must be replaced with bar5,
+        // thereofore the returned list must have a size of 2
+        final List<Bar> replacedBars = BarSeriesUtils.replaceBarsIfChanged(this.series, newBars);
+
+        // the replaced bar must be the same as the previous bar
+        assertEquals(2, replacedBars.size());
+    }
+
     @Test
     public void findMissingBarsTest() {